Extract class into separate file
