+++ app/src/main/java/com/marverenic/music/utils/RxProperty.java
@@ -1,0 +1,75 @@
+package com.marverenic.music.utils;
+
+import rx.Observable;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
+import rx.subjects.BehaviorSubject;
+import timber.log.Timber;
+
+public final class RxProperty<T> {
+
+    private final String mName;
+    private final T mNullValue;
+    private final BehaviorSubject<Optional<T>> mSubject;
+    private final Observable<T> mObservable;
+
+    private Retriever<T> mRetriever;
+
+    public RxProperty(String propertyName) {
+        this(propertyName, null);
+    }
+
+    public RxProperty(String propertyName, T nullValue) {
+        mName = propertyName;
+        mNullValue = nullValue;
+        mSubject = BehaviorSubject.create();
+
+        mObservable = mSubject.filter(Optional::isPresent)
+                .map(Optional::getValue)
+                .distinctUntilChanged();
+    }
+
+    public void setFunction(Retriever<T> retriever) {
+        mRetriever = retriever;
+    }
+
+    public void invalidate() {
+        mSubject.onNext(Optional.empty());
+
+        if (mRetriever != null) {
+            Observable.fromCallable(mRetriever::retrieve)
+                    .subscribeOn(Schedulers.computation())
+                    .map(data -> (data == null) ? mNullValue : data)
+                    .map(Optional::ofNullable)
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(mSubject::onNext, throwable -> {
+                        Timber.e(throwable, "Failed to fetch " + mName + " property.");
+                    });
+        }
+    }
+
+    public boolean isSubscribedTo() {
+        return mSubject.hasObservers();
+    }
+
+    public void setValue(T value) {
+        mSubject.onNext(Optional.ofNullable(value));
+    }
+
+    public boolean hasValue() {
+        return mSubject.getValue() != null && mSubject.getValue().isPresent();
+    }
+
+    public T lastValue() {
+        return mSubject.getValue().getValue();
+    }
+
+    public Observable<T> getObservable() {
+        return mObservable;
+    }
+
+    public interface Retriever<T> {
+        T retrieve() throws Exception;
+    }
+
+}
